make gql for gatekeeper more modular
