+++ src/main/java/org/icgc/argo/workflow_management/graphql/GraphQLProvider.java
@@ -23,27 +23,27 @@ import static graphql.schema.idl.TypeRuntimeWiring.newTypeWiring;
 import com.apollographql.federation.graphqljava.Federation;
 import com.google.common.base.Charsets;
 import com.google.common.io.Resources;
+import com.sun.xml.bind.v2.TODO;
 import graphql.GraphQL;
 import graphql.scalars.ExtendedScalars;
+import graphql.schema.DataFetcher;
 import graphql.schema.GraphQLSchema;
 import graphql.schema.idl.RuntimeWiring;
 import java.io.IOException;
 import java.net.URL;
 import javax.annotation.PostConstruct;
-import org.springframework.beans.factory.annotation.Autowired;
+
+import graphql.schema.visibility.BlockedFields;
+import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.stereotype.Service;
 
 @Service
+@RequiredArgsConstructor
 public class GraphQLProvider {
   private final MutationDataFetcher mutationDataFetcher;
+  private final DataFetcher gatekeeperDataFetcher;
   private GraphQL graphQL;
-  private GraphQLSchema graphQLSchema;
-
-  @Autowired
-  public GraphQLProvider(MutationDataFetcher mutationDataFetcher) {
-    this.mutationDataFetcher = mutationDataFetcher;
-  }
 
   @Bean
   public GraphQL graphQL() {
@@ -54,7 +54,7 @@ public class GraphQLProvider {
   public void init() throws IOException {
     URL url = Resources.getResource("schema.graphql");
     String sdl = Resources.toString(url, Charsets.UTF_8);
-    graphQLSchema = buildSchema(sdl);
+    GraphQLSchema graphQLSchema = buildSchema(sdl);
     this.graphQL = GraphQL.newGraphQL(graphQLSchema).build();
   }
 
@@ -66,7 +66,10 @@ public class GraphQLProvider {
     return RuntimeWiring.newRuntimeWiring()
         .scalar(ExtendedScalars.Json)
         .type(newTypeWiring("Mutation").dataFetchers(mutationDataFetcher.mutationResolvers()))
-        .fieldVisibility(new AdminVisibility())
+        .type(newTypeWiring("Query").dataFetcher("activeRuns", gatekeeperDataFetcher))
+        .fieldVisibility(BlockedFields.newBlock()
+                                 .addPattern("activeRuns") // TODO make visible after api is removed
+                                 .build())
         .build();
   }
 }
